move get_config function
