+++ text_clf/config.py
@@ -1,6 +1,55 @@
+import ast
+import datetime
 import logging
 import os
 import sys
+from pathlib import Path
+from typing import Any, Dict
+
+import yaml
+
+
+def get_config(path_to_config: str) -> Dict[str, Any]:
+    """
+    Get config.
+
+    :param str path_to_config: path to config.
+    :return: config.
+    :rtype: Dict[str, Any]
+    """
+
+    if not os.path.exists(path_to_config):
+        load_default_config()
+
+    with open(path_to_config, mode="r") as fp:
+        config = yaml.safe_load(fp)
+
+    now = datetime.datetime.now()
+
+    config["path_to_save_folder"] = (
+        Path(config["path_to_save_folder"]) / f"model_{now.date()}_{now.time()}"
+    )
+
+    config["path_to_config"] = path_to_config
+    config["path_to_save_model"] = config["path_to_save_folder"] / "model.joblib"
+    config["path_to_save_logfile"] = config["path_to_save_folder"] / "logging.txt"
+    config["path_to_save_target_names_mapping"] = (
+        config["path_to_save_folder"] / "target_names.json"
+    )
+
+    # tf-idf
+    if ("tf-idf" not in config) or (config["tf-idf"] is None):
+        config["tf-idf"] = {}
+    if "ngram_range" in config["tf-idf"]:
+        config["tf-idf"]["ngram_range"] = ast.literal_eval(
+            config["tf-idf"]["ngram_range"]
+        )
+
+    # logreg
+    if ("logreg" not in config) or (config["logreg"] is None):
+        config["logreg"] = {}
+
+    return config
 
 
 def get_logger() -> logging.Logger:
